+++ System/Metrics/Gauge.hs
@@ -15,6 +15,7 @@ module System.Metrics.Gauge
     ) where
 
 import qualified Data.Atomic as Atomic
+import Data.Int (Int64)
 import Prelude hiding (subtract, read)
 
 -- | A mutable, integer-valued gauge.
@@ -25,7 +26,7 @@ new :: IO Gauge
 new = C `fmap` Atomic.new 0
 
 -- | Get the current value of the gauge.
-read :: Gauge -> IO Int
+read :: Gauge -> IO Int64
 read = Atomic.read . unC
 
 -- | Increase the gauge by one.
@@ -37,13 +38,13 @@ dec :: Gauge -> IO ()
 dec gauge = subtract gauge 1
 
 -- | Increase the gauge by the given amount.
-add :: Gauge -> Int -> IO ()
+add :: Gauge -> Int64 -> IO ()
 add gauge = Atomic.add (unC gauge)
 
 -- | Decrease the gauge by the given amount.
-subtract :: Gauge -> Int -> IO ()
+subtract :: Gauge -> Int64 -> IO ()
 subtract gauge = Atomic.subtract (unC gauge)
 
 -- | Set the gauge to the given value.
-set :: Gauge -> Int -> IO ()
+set :: Gauge -> Int64 -> IO ()
 set gauge = Atomic.write (unC gauge)
